balance formatting

+++ app/src/main/java/com/balance/boi/database/data/Balance.kt
@@ -3,6 +3,7 @@ package com.balance.boi.database.data
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import androidx.room.TypeConverters
+import java.math.BigDecimal
 import java.sql.Date
 
 @Entity
@@ -11,5 +12,5 @@ data class Balance(
     @PrimaryKey(autoGenerate = true) val balanceId: Int = 0,
     val balanceAccountId: Int,
     val balanceDate: Date,
-    val balanceAmount: Double
+    val balanceAmount: BigDecimal
 )
